+++ Cargo.toml
@@ -50,6 +50,7 @@ openat-ext = "^0.1.13"
 os-release = "0.1.0"
 ostree = { version = "0.9.1", features = ["v2020_4"] }
 ostree-sys = "0.7.1"
+ostree-host = { path = "rust/ostree-host" }
 paste = "1.0"
 phf = { version = "0.8", features = ["macros"] }
 rand = "0.8.3"
